add favorite feature

+++ src/main/java/com/example/miniweibo/controller/CommentController.java
@@ -22,24 +22,38 @@ public class CommentController {
     @PostMapping("/addComment")
     public Comment addComment(@RequestBody AddCommentRequest addCommentRequest, HttpSession httpSession){
         Long userId = (Long) httpSession.getAttribute("userId");
+        if(userId == null){
+            throw new RuntimeException("ログインしてください。");
+        }
         User user = userRepository.findById(userId).get();
         return commentService.addComment(user,addCommentRequest.getWeiboId(),addCommentRequest.getContent());
     }
     @PostMapping("/deleteComment")
     public void deleteComment(HttpSession httpSession,@RequestParam Long commentId){
         Long userId = (Long) httpSession.getAttribute("userId");
+        if(userId == null){
+            throw new RuntimeException("ログインしてください。");
+        }
         User user = userRepository.findById(userId).get();
         commentService.deleteComment(user,commentId);
     }
-    @PostMapping("/updateComment")
-    public Comment updateComment(HttpSession httpSession,@RequestParam Long commentId,
+    @PutMapping("/comment/{commentId}")
+    public Comment updateComment(HttpSession httpSession,
+                                 @PathVariable Long commentId,
                                  @RequestBody AddCommentRequest addCommentRequest){
         Long userId = (Long) httpSession.getAttribute("userId");
-        User user = userRepository.findById(userId).get();
+        if(userId == null){
+            throw new RuntimeException("ログインしてください。");
+        }
+        User user = userRepository.findById(userId).orElseThrow();
         return commentService.updateComment(user,commentId,addCommentRequest.getContent());
     }
     @GetMapping("/getCommentsByWeibo")
-    public List<Comment> getCommentsByWeibo(@RequestParam Long weiboId){
+    public List<Comment> getCommentsByWeibo(@RequestParam Long weiboId,HttpSession httpSession){
+        Long userId = (Long) httpSession.getAttribute("userId");
+        if (userId == null) {
+            throw new RuntimeException("ログインしてください。");
+        }
         return commentService.getCommentsByWeibo(weiboId);
     }
 }
